Move KnifeTrace and KnifeSmack back to CKnife
They don't really belong in MomUtil, and cleaned them up a bit
