+++ mp/src/game/shared/momentum/util/mom_util.cpp
@@ -464,160 +464,6 @@ void MomUtil::KVLoadQAngles(KeyValues *kvFrom, const char *pName, QAngle &angInt
     LOAD_3D_FROM_KV(kvFrom, pName, angInto);
 }
 
-inline void FindHullIntersection(const Vector &vecSrc, trace_t &tr, const Vector &mins, const Vector &maxs, CBaseEntity *pEntity)
-{
-    int     i, j, k;
-    float   distance;
-    Vector  minmaxs[2] = { mins, maxs };
-    trace_t tmpTrace;
-    Vector  vecHullEnd = tr.endpos;
-    Vector  vecEnd;
-
-    distance = 1e6f;
-
-    vecHullEnd = vecSrc + ((vecHullEnd - vecSrc) * 2);
-    UTIL_TraceLine(vecSrc, vecHullEnd, MASK_SOLID, pEntity, COLLISION_GROUP_NONE, &tmpTrace);
-    if (tmpTrace.fraction < 1.0)
-    {
-        tr = tmpTrace;
-        return;
-    }
-
-    for (i = 0; i < 2; i++)
-    {
-        for (j = 0; j < 2; j++)
-        {
-            for (k = 0; k < 2; k++)
-            {
-                vecEnd.x = vecHullEnd.x + minmaxs[i][0];
-                vecEnd.y = vecHullEnd.y + minmaxs[j][1];
-                vecEnd.z = vecHullEnd.z + minmaxs[k][2];
-
-                UTIL_TraceLine(vecSrc, vecEnd, MASK_SOLID, pEntity, COLLISION_GROUP_NONE, &tmpTrace);
-                if (tmpTrace.fraction < 1.0)
-                {
-                    float thisDistance = (tmpTrace.endpos - vecSrc).Length();
-                    if (thisDistance < distance)
-                    {
-                        tr = tmpTrace;
-                        distance = thisDistance;
-                    }
-                }
-            }
-        }
-    }
-}
-
-void MomUtil::KnifeTrace(const Vector& vecShootPos, const QAngle& lookAng, bool bStab, CBaseEntity *pAttacker,
-    CBaseEntity *pSoundSource, trace_t* trOutput, Vector* vForwardOut)
-{
-    float fRange = bStab ? 32.0f : 48.0f; // knife range
-
-    AngleVectors(lookAng, vForwardOut);
-    Vector vecSrc = vecShootPos;
-    Vector vecEnd = vecSrc + *vForwardOut * fRange;
-
-    UTIL_TraceLine(vecSrc, vecEnd, MASK_SOLID, pAttacker, COLLISION_GROUP_NONE, trOutput);
-
-#ifndef CLIENT_DLL
-    if (pAttacker->IsPlayer())
-    {
-        //check for hitting glass
-        CTakeDamageInfo glassDamage(pAttacker, pAttacker, 42.0f, DMG_BULLET | DMG_NEVERGIB);
-        pSoundSource->TraceAttackToTriggers(glassDamage, trOutput->startpos, trOutput->endpos, *vForwardOut);
-    }
-#endif
-
-    if (trOutput->fraction >= 1.0f)
-    {
-        Vector head_hull_mins(-16, -16, -18);
-        Vector head_hull_maxs(16, 16, 18);
-        UTIL_TraceHull(vecSrc, vecEnd, head_hull_mins, head_hull_maxs, MASK_SOLID, pAttacker, COLLISION_GROUP_NONE, trOutput);
-        if (trOutput->fraction < 1.0f)
-        {
-            // Calculate the point of intersection of the line (or hull) and the object we hit
-            // This is and approximation of the "best" intersection
-            CBaseEntity *pHit = trOutput->m_pEnt;
-            if (!pHit || pHit->IsBSPModel())
-                FindHullIntersection(vecSrc, *trOutput, VEC_DUCK_HULL_MIN, VEC_DUCK_HULL_MAX, pAttacker);
-            //vecEnd = trOutput->endpos; // This is the point on the actual surface (the hull could have hit space)
-        }
-    }
-
-    bool bDidHit = trOutput->fraction < 1.0f;
-
-    if (!bDidHit)
-    {
-        // play wiff or swish sound
-        CPASAttenuationFilter filter(trOutput->endpos, "Weapon_Knife.Slash");
-        filter.UsePredictionRules();
-        CBaseEntity::EmitSound(filter, pSoundSource->entindex(), "Weapon_Knife.Slash");
-    }
-#ifndef CLIENT_DLL
-    else if (pAttacker->IsPlayer())
-    {
-        CMomentumPlayer *pPlayer = static_cast<CMomentumPlayer*>(pAttacker);
-        if (pPlayer->m_bHasPracticeMode)
-            return;
-
-        CBaseEntity *pHitEntity = trOutput->m_pEnt;
-
-        ClearMultiDamage();
-
-        const float flDamage = bStab ? 65.0f : 20.0f;
-
-        CTakeDamageInfo info( pAttacker, pAttacker, flDamage, DMG_BULLET | DMG_NEVERGIB );
-
-        CalculateMeleeDamageForce( &info, *vForwardOut, trOutput->endpos, 1.0f/flDamage );
-        pHitEntity->DispatchTraceAttack(info, *vForwardOut, trOutput); 
-        ApplyMultiDamage();
-    }
-#endif
-}
-
-void MomUtil::KnifeSmack(const trace_t& trIn, CBaseEntity *pSoundSource, const QAngle& lookAng, const bool bStab)
-{
-    if (!trIn.m_pEnt || (trIn.surface.flags & SURF_SKY))
-        return;
-
-    if (trIn.fraction == 1.0f)
-        return;
-
-    if (trIn.m_pEnt)
-    {
-        CPASAttenuationFilter filter(trIn.endpos);
-        filter.UsePredictionRules();
-
-        const auto pRunEnt = dynamic_cast<CMomRunEntity*>(trIn.m_pEnt);
-        if (pRunEnt)
-        {
-            CBaseEntity::EmitSound(filter, pSoundSource->entindex(), bStab ? "Weapon_Knife.Stab" : "Weapon_Knife.Hit");
-        }
-        else
-        {
-            CBaseEntity::EmitSound(filter, pSoundSource->entindex(), "Weapon_Knife.HitWall");
-        }
-    }
-
-    CEffectData data;
-    data.m_vOrigin = trIn.endpos;
-    data.m_vStart = trIn.startpos;
-    data.m_nSurfaceProp = trIn.surface.surfaceProps;
-    data.m_nDamageType = DMG_SLASH;
-    data.m_nHitBox = trIn.hitbox;
-#ifdef CLIENT_DLL
-    data.m_hEntity = trIn.m_pEnt->GetRefEHandle();
-#else
-    data.m_nEntIndex = trIn.m_pEnt->entindex();
-#endif
-
-    CPASFilter filter(data.m_vOrigin);
-    data.m_vAngles = lookAng;
-    data.m_fFlags = 0x1; //IMPACT_NODECAL;
-
-    te->DispatchEffect(filter, 0.0, data.m_vOrigin, "KnifeSlash", data);
-}
-
 bool MomUtil::GetSHA1Hash(const CUtlBuffer& buf, char* pOut, size_t outLen)
 {
     CryptoPP::SHA1 hash;
